pgbloomfun: add automatic scaling of the bloom filters
Scalable filters are implemented by creating new (larger) filters as the old
ones are filled; the growth factor is configurable but error rate is not
adjusted for the new filters, meaning that the total error rate will
increase as more filters are added.

Also added a function for retrieving statistics on the filters.