pgbloomfun: support for postgresql versions before 9.2
Extensions are only supported in PostgreSQL 9.1 and newer version, this
commit allows building and installing pgbloomfun without extension support
(for example with PG 8.4) and adds an upgrade script to upgrade the schema
to an extension after a PostgreSQL upgrade.  Also switched bloom_stats()
output from JSON to TEXT in PostgreSQL < 9.2.